fix: drop a redundant u64 cast in the new regression test
Clippy runs with `-D warnings` over all targets, where `(state >> 33) as u64`
on a `u64` trips `unnecessary_cast`.

+++ datafusion/core/tests/memory_limit/nlj_spill_unmatched.rs
@@ -38,7 +38,7 @@ fn table(rows: usize, value_column: &str, seed: u64) -> RecordBatch {
         state = state
             .wrapping_mul(6364136223846793005)
             .wrapping_add(1442695040888963407);
-        (state >> 33) as u64
+        state >> 33
     };
     let schema = Arc::new(Schema::new(vec![
         Field::new("k", DataType::Int32, true),
